requirements.txt
discord.py
wavelink
requests
aiohttp
google-genai
PyNaCl